Move 'show platform psustatus' tests to show_platform_test.py
